refactor: move configuration to cli package
