+++ hexagon/cli/__init__.py
@@ -1,0 +1,11 @@
+import os
+
+from cli.configuration import Configuration
+
+CONFIG_FILE_ENV_VARIABLE_NAME = "HEXAGON_CONFIG_FILE"
+
+
+configuration = Configuration()
+cli, tools, envs = configuration.init_config(
+    os.getenv(CONFIG_FILE_ENV_VARIABLE_NAME, "app.yaml")
+)
