+++ hexagon/support/execute_tool.py
@@ -4,7 +4,7 @@ import sys
 import os
 from typing import List, Union, Dict
 
-from hexagon.support.config import configuration
+from hexagon.cli import configuration
 from hexagon.support.printer import log
 
 _command_by_file_extension = {"js": "node", "sh": "bash"}
