+++ hexagon/tools/internal/install_cli.py
@@ -4,7 +4,7 @@ from InquirerPy import inquirer
 from InquirerPy.validator import PathValidator
 from prompt_toolkit.validation import ValidationError
 
-from hexagon.support.config import configuration
+from hexagon.cli import configuration
 from hexagon.tools.internal.save_new_alias import save_new_alias
